Makefile: ASAN rules for Rust lib and example
ASAN (Address SANitizer) is a tool that detects memory corruption bugs.
This may be extremely useful when working with unsafe code in Rust
and trying to understand where the memory corruption happens.
With this Makefile target, you can build the Rust library and the C#
example with ASAN enabled, enjoying the insights provided by ASAN's
runtime checks.