C#-Rust interop PoC
This is a PoC of interop between C# and Rust via FFI.
List of features supported so far:
- Creating a Session,
- Executing a simple query,
- Deserializing query's results into RowSet and Row,
- Preparing a statement.

The binding layer uses bare FFI on both sides (C# and Rust).
It's painfully raw, especially on the C# side, but it works.
Let's get inspired by this PoC, and build a more user-friendly
binding layer in the future.
The goal of this PoC is to present the idea, and to show that
it works with not too much effort. It also serves as a reference
for setting the FFI to call Rust from C# and vice versa.